refactor(football): extract dataclasses to bot/domain/models/football
Move TmPlayer, TmClub, TmSquadStats, SportsDBTeam, and StandingRow
from service files to a shared domain models module, establishing
clean dependency direction (services depend on models, not vice-versa).
