+++ bot/domain/commands/football_team.py
@@ -25,15 +25,11 @@ from bot.domain.commands.base import (
     Platform,
 )
 from bot.domain.models.command_data import CommandData
+from bot.domain.models.football import SportsDBTeam, TmClub, TmPlayer, TmSquadStats
 from bot.domain.models.message import BotMessage
 from bot.domain.services.football_field_builder import build_football_field
-from bot.domain.services.thesportsdb import SportsDBTeam, TheSportsDBService
-from bot.domain.services.transfermarkt import (
-    TmClub,
-    TmPlayer,
-    TmSquadStats,
-    TransfermarktService,
-)
+from bot.domain.services.thesportsdb import TheSportsDBService
+from bot.domain.services.transfermarkt import TransfermarktService
 from bot.infrastructure.http_client import HttpClient
 
 logger = structlog.get_logger()
