+++ tests/unit/commands/test_football_player.py
@@ -2,8 +2,9 @@ import pytest
 
 from bot.data.football import LEAGUES
 from bot.domain.commands.football_player import FootballPlayerCommand
+from bot.domain.models.football import TmPlayer
 from bot.domain.models.message import ImageBufferContent
-from bot.domain.services.transfermarkt import TmPlayer, TransfermarktService
+from bot.domain.services.transfermarkt import TransfermarktService
 from bot.infrastructure.http_client import HttpClient
 from tests.factories.command_data import GroupCommandDataFactory, PrivateCommandDataFactory
